Created using spr 1.3.6-beta.1

+++ wicket/src/wicketd.rs
@@ -28,13 +28,13 @@ impl From<ComponentId> for SpIdentifier {
     fn from(id: ComponentId) -> Self {
         match id {
             ComponentId::Sled(i) => {
-                SpIdentifier { type_: SpType::Sled, slot: u32::from(i) }
+                SpIdentifier { type_: SpType::Sled, slot: u16::from(i) }
             }
             ComponentId::Psc(i) => {
-                SpIdentifier { type_: SpType::Power, slot: u32::from(i) }
+                SpIdentifier { type_: SpType::Power, slot: u16::from(i) }
             }
             ComponentId::Switch(i) => {
-                SpIdentifier { type_: SpType::Switch, slot: u32::from(i) }
+                SpIdentifier { type_: SpType::Switch, slot: u16::from(i) }
             }
         }
     }
